refactor: refactored build script
chore: small changes

1. removed chinese comments
2. fixed bash script error

+++ build.sh
@@ -1,14 +1,15 @@
 cd react
 npm install
 npm run build
-rm -r ../public/markdown-viewer
+[ -d ../public/markdown-viewer ] && rm -r ../public/markdown-viewer
 mkdir -p ../public/markdown-viewer
 cp -r dist/* ../public/markdown-viewer/
 
 cd ..
 go build
 mkdir -p dist
-cp simple-server dist/
+[ -f simple-server ] && cp simple-server dist/
+[ -f simple-server.exe ] && cp simple-server.exe dist/
 [ -f config.yaml ] && cp config.yaml dist/
-cp config.yaml.example dist/
+[ -f config.yaml.example ] && cp config.yaml.example dist/
 cp -r public dist/
